Move to a fork of curve25519-dalek, dalek-ff-group to use a Field25519 from curve25519-dalek
Specifically, this uses the fiat-crypto backend for curve25519-dalek which has
been formally verified. Notably, since we're now resolving to solely a single
backend, we could alternatively not use a patch of curve25519-dalek yet add a
direct dependency of fiat-crypto to the dalek-ff-group fork to achieve (almost)
the same result. Currently, we'll only benefit from curve25519-dalek providing
some more tooling around fiat-crypto and switching between the 32- and 64-bit
variants.

This has been tested on 64-bit platforms. This has yet to be tested on 32-bit
platforms.